requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
jinja2==3.1.4
python-multipart==0.0.9
python-escpos==3.1
